requirements.txt
fastapi
uvicorn[standard]
requests
numpy
pandas
faiss-cpu
sentence-transformers
python-dotenv
tqdm